chore(examples): add weather demo app
