+++ examples/apps/weather/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "weather"
+publish = false
+license.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+color-eyre.workspace = true
+crossterm.workspace = true
+rand = "0.8.5"
+ratatui.workspace = true
+
+[lints]
+workspace = true
