+++ ratatui/Cargo.toml
@@ -143,11 +143,6 @@ name = "main"
 harness = false
 
 [[example]]
-name = "barchart"
-required-features = ["crossterm"]
-doc-scrape-examples = true
-
-[[example]]
 name = "barchart-grouped"
 required-features = ["crossterm"]
 doc-scrape-examples = true
